@@ -1,5 +1,5 @@
 from pyiron_workflow import as_function_node
-from node_library.atomistic.engine.generic import OutputEngine
+from pyiron_nodes.atomistic.engine.generic import OutputEngine
 
 
 @as_function_node("engine")
@@ -16,7 +16,9 @@ def M3GNet():
     import matgl
     from matgl.ext.ase import M3GNetCalculator
 
-    out = OutputEngine(calculator=M3GNetCalculator(matgl.load_model("M3GNet-MP-2021.2.8-PES")))
+    out = OutputEngine(
+        calculator=M3GNetCalculator(matgl.load_model("M3GNet-MP-2021.2.8-PES"))
+    )
     return out
 
 
