@@ -1,9 +1,9 @@
 import unittest
-import pyiron_module_template
+import pyiron_nodes
 
 
 class TestVersion(unittest.TestCase):
     def test_version(self):
-        version = pyiron_module_template.__version__
+        version = pyiron_nodes.__version__
         print(version)
         self.assertTrue(version.startswith('0'))
